+++ presentation-compose/src/main/kotlin/me/proton/core/compose/component/ProtonCenteredProgress.kt
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2021 Proton Technologies AG
+ * This file is part of Proton Technologies AG and ProtonCore.
+ *
+ * ProtonCore is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * ProtonCore is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with ProtonMail.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package me.proton.core.compose.component
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun ProtonCenteredProgress(
+    modifier: Modifier = Modifier
+) {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = modifier
+    ) {
+        CircularProgressIndicator()
+    }
+}
+
+@Preview(
+    heightDp = 100,
+    widthDp = 100,
+    showBackground = true,
+    backgroundColor = android.graphics.Color.WHITE.toLong()
+)
+@Composable
+fun PreviewProtonCenteredProgress() {
+    ProtonCenteredProgress()
+}