+++ presentation-compose/src/main/kotlin/me/proton/core/compose/component/VerticalSpacer.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 Proton Technologies AG
+ * Copyright (c) 2021 Proton Technologies AG
  * This file is part of Proton Technologies AG and ProtonCore.
  *
  * ProtonCore is free software: you can redistribute it and/or modify
@@ -7,7 +7,7 @@
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * ProtonCore is distributed in the hope that it will be useful,
+ * ProtonMail is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
@@ -16,12 +16,19 @@
  * along with ProtonCore.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import org.gradle.api.Project
-import org.gradle.kotlin.dsl.repositories
+package me.proton.core.compose.component
 
-internal fun Project.applyRepositories() {
-    repositories {
-        google()
-        mavenCentral()
-    }
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.Dp
+import me.proton.core.compose.theme.ProtonDimens
+
+@Composable
+fun VerticalSpacer(
+    modifier: Modifier = Modifier,
+    height: Dp = ProtonDimens.SmallSpacing
+) {
+    Spacer(modifier.height(height))
 }